# py3/app/views/models.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml
#   timestamp: 2023-01-28T19:34:24+00:00

from __future__ import annotations

from typing import Optional

from pydantic import BaseModel, EmailStr, Field, SecretStr


class User(BaseModel):
    id: Optional[str] = Field(None, description='Идентификатор пользователя')
    email: Optional[EmailStr] = Field(
        None, description='Адрес электронной почты', example='ivan@email.com'
    )
    first_name: Optional[str] = Field(None, description='Имя', example='Иван')
    last_name: Optional[str] = Field(None, description='Фамилия', example='Иванов')
    age: Optional[int] = Field(None, description='Возраст', example=18)
    biography: Optional[str] = Field(
        None, description='Интересы', example='Хобби, интересы и т.п.'
    )
    city: Optional[str] = Field(None, description='Город', example='Москва')
    password: Optional[SecretStr] = Field(None, example='qwerty')


class ListUsersRequest(BaseModel):
    first_name: Optional[str] = Field(
        None, description='Условие поиска по имени', example='Конст'
    )
    last_name: Optional[str] = Field(
        None, description='Условие поиска по фамилии', example='Оси'
    )
    offset: Optional[int] = Field(0, description='Начало страницы результата')
    limit: Optional[int] = Field(20, description='Размер страницы результата')


class AuthLoginPostRequest(BaseModel):
    email: Optional[str] = Field(None, example='ivan@email.com')
    password: Optional[SecretStr] = Field(None, example='qwerty')


class AuthLoginPostResponse(BaseModel):
    token: Optional[str] = Field(None, example='e4d2e6b0cde242c5aac30b8316f21e58')